OS/2 personality enhancements - Unify RPC interfaces of os2srv. os2exec and os2fs. Now all three expose a single "CPI" service. In Genode, session requests with label "exec" go to os2exec, session requests with label "fs" go to os2fs, and all other ones are handled by os2srv. - Use u32_t 32-bit unsigned type, defined to "unsigned int" on 64-bit arch's and to "unsigned long" on 32-bit and 16-bit arch's. (only used in exe386.h header so far). - Separate a platform-independent init code from platform-dependent one in all three servers and os2app. - Add some build system for L4Re case.

+++ OS2/Server/include/os2/os2def.h
@@ -60,7 +60,8 @@ typedef unsigned long   ULONG, *PULONG;
 typedef USHORT WORD;
 # endif
 # ifndef DWORD
-typedef ULONG  DWORD;
+//typedef ULONG  DWORD;
+typedef UINT  DWORD;
 # endif
 #endif
